chips: rp2040: move the PIO driver into the shared rp2xxx crate
The RP2350 has the same PIO hardware and chips/rp2350 has no driver for
it. Move this one alongside the PL022 SPI driver rather than writing a
second copy.

The two chips lay the block out identically from CTRL through the per
state machine registers, which is everything this driver touches. They
differ in where the interrupt registers start: the RP2040 puts them
directly after the state machines at +0x128, and the RP2350 inserts
sixteen RXFn_PUTGETn registers and a GPIOBASE first, so they begin at
+0x16c. That is an address rather than a layout, so PioRegisters now
ends at 0x128 and the interrupt registers are a second block whose base
each chip supplies. The four state machines per block and thirty two
instruction words the driver hardcodes hold on both chips: an RP2350
reads DBG_CFGINFO 0x10200404, giving SM_COUNT 4 and IMEM_SIZE 32.

Both chip types stay concrete, following spi.rs. Pio is generic over the
chip's own block type, and rp2040 supplies
`pub type Pio = rp2xxx::pio::Pio<PIONumber>;`, so PIONumber stays in this
crate and DmaPeripheral::PioRxFifo keeps taking one. Pins arrive through
a one-method trait, PioPin, for the same reason PeripheralClock exists:
the driver needs a pin number for a register field and kernel::hil::gpio
has none. So set_pins and set_input_sync_bypass take `&[&P]` and `&P`,
and RPGpioPin::pin() being pub(crate) is no obstacle since the impl
lives here. PioBlock itself has no methods; the driver only carries a
block so number() can hand it back.

gpio_init does not move. GpioFunction is not the same enum on the two
chips, 8 being GPCK here and PIO2 there, so it becomes a free function
taking the Pio it points a pin at.

One thing is not a pure relocation. rx_fifo_addr and tx_fifo_addr took a
PIONumber and recomputed an address from a hardcoded base, alongside a
state machine that already knew its block, so the two could disagree.
They now read it from the registers they hold, as
chips/msp432/src/uart.rs does: txf is at +0x010 and rxf at +0x020, four
u32 each, which is the same address the arithmetic produced.

chips/rp2040/src/pio.rs is 97 lines now. Boards still write
rp2040::pio::Pio; the one caller visible change is Pio::new_pio0()
becoming pio::new_pio0(), as the SPI move did, touching one line in
chip.rs and one in boards/pico_explorer_base.

    board                 text          bss
    raspberry_pi_pico       +0   16528 -> 16568
    pico_explorer_base      +0   72668 -> 72708
    nano_rp2040_connect     +0   14368 -> 14408
    raspberry_pi_pico_w     +0   20724 -> 20764

text is unchanged on all four boards. bss grows 40 bytes on each for the
same reason: Pio and its four state machines now hold a pointer to the
interrupt block instead of computing its address, 4 bytes across 2
blocks of 5 objects.

cargo fmt --check, tools/ci/check_format.sh and cargo clippy are clean.
Nothing in this commit was run on RP2040 hardware; the DBG_CFGINFO
reading is from an RP2350 and predates it.
