+++ chips/rp2040/src/chip.rs
@@ -157,8 +157,8 @@ impl Rp2040DefaultPeripherals<'_> {
             dma: dma::Dma::new(),
             i2c0: i2c::I2c::new_i2c0(clocks, resets),
             pins: RPPins::new(),
-            pio0: Pio::new_pio0(),
-            pio1: Pio::new_pio1(),
+            pio0: crate::pio::new_pio0(),
+            pio1: crate::pio::new_pio1(),
             pwm: pwm::Pwm::new(clocks),
             sio: SIO::new(),
             spi0: spi::new_spi0(clocks),
