+++ chips/rp2040/src/pio_gspi.rs
@@ -157,8 +157,7 @@ impl PioGSpi<'_> {
     fn dma_pull(&self, addr: u32, len: u32) {
         assert!(addr.is_multiple_of(4));
         let current_sm = self.pio.sm(self.sm_number);
-        self.dma
-            .set_read_addr(current_sm.rx_fifo_addr(self.pio.number()));
+        self.dma.set_read_addr(current_sm.rx_fifo_addr());
         self.dma.set_write_addr(addr);
 
         self.dma.set_len(len);
@@ -174,8 +173,7 @@ impl PioGSpi<'_> {
         assert!(addr.is_multiple_of(4));
         let current_sm = self.pio.sm(self.sm_number);
         self.dma.set_read_addr(addr);
-        self.dma
-            .set_write_addr(current_sm.tx_fifo_addr(self.pio.number()));
+        self.dma.set_write_addr(current_sm.tx_fifo_addr());
 
         self.dma.set_len(len);
 
@@ -294,8 +292,8 @@ fn cyw43_spi_program_init(
         RPGpioPin::new(RPGpio::from_u32(clock_pin).expect("GPIO pin must be 0 to 29"));
     let dio_pin_handle =
         RPGpioPin::new(RPGpio::from_u32(dio_pin).expect("GPIO pin must be 0 to 29"));
-    pio.gpio_init(&clock_pin_handle);
-    pio.gpio_init(&dio_pin_handle);
+    crate::pio::gpio_init(pio, &clock_pin_handle);
+    crate::pio::gpio_init(pio, &dio_pin_handle);
 
     dio_pin_handle.set_floating_state(kernel::hil::gpio::FloatingState::PullNone);
     dio_pin_handle.set_schmitt(true);
