+++ chips/rp2040/src/pio_pwm.rs
@@ -114,9 +114,10 @@ fn pwm_program_init(
     // "out isr, 32" command created by pioasm
     let out_isr_32_command = 0x60c0_u16;
     sm.config(config);
-    pio.gpio_init(&RPGpioPin::new(
-        RPGpio::from_u32(pin).expect("GPIO pin must be 0 to 29"),
-    ));
+    crate::pio::gpio_init(
+        pio,
+        &RPGpioPin::new(RPGpio::from_u32(pin).expect("GPIO pin must be 0 to 29")),
+    );
     sm.set_enabled(false);
     sm.set_pins_dirs(pin, 1, true);
     sm.set_side_set_pins(pin, 1, false, true);
